refact :  button variants
