+++ client/src/shared/ui/button/index.ts
@@ -1,2 +1,1 @@
 export { default as MainButton } from './MainButton';
-export * as BtnType from './type';
